Server/requirements/development.txt
-r base.txt
django==2.0.4
django-debug-toolbar==1.9.1
pytz==2018.3
sqlparse==0.2.4
werkzeug==0.14.1
dj-static==0.0.6
aiohttp==3.4.4